# public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"ApiResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {}
				},
				"required": [
					"success",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Status": {
				"enum": [
					"ACTIVE",
					"INACTIVE",
					"DEACTIVATED"
				],
				"type": "string"
			},
...
			"CreateCollegeDTO": {
				"properties": {
					"name": {
						"type": "string"
					},
					"code": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"contactPhone": {
						"type": "string"
					},
					"contactEmail": {
						"type": "string"
					},
					"adminName": {
						"type": "string"
					},
					"adminEmail": {
						"type": "string"
					},
					"adminPassword": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"code"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateCollegeDTO": {
				"properties": {
					"name": {
						"type": "string"
					},
					"code": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"contactPhone": {
						"type": "string"
					},
					"contactEmail": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/Status"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "access_token",
				"in": "cookie"
			}
		}
	},
	"info": {
		"title": "busapp-backend",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/api/v1/colleges": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "List all colleges on the platform.\n\n### Intent & Business Purpose\nReturns a list of active college tenants registered on the bus tracking platform.\n\n### Target Audience & Roles\n- **Allowed Roles:** Authenticated users (`ADMIN`, `COLLEGE_ADMIN`, `DRIVER`, `PARENT`, `STUDENT`).",
				"tags": [
					"Colleges"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Register a new college tenant.\n\n### Intent & Business Purpose\nCreates a new college organization entity on the platform. Enables multi-tenancy for the new institution.\n\n### Target Audience & Roles\n- **Allowed Roles:** System Administrator (`ADMIN`).\n\n### Request Body\n- `name` *(required string)*: Full college name (e.g. `Kathmandu University`).\n- `code` *(required string)*: Unique short code (e.g. `KU`).\n- `address` *(optional string)*: Physical address.\n- `contactEmail` *(optional string)*: Primary administrative email.\n\n### Side Effects\n- Inserts a new row into the `colleges` table.",
				"tags": [
					"Colleges"
				],
				"security": [
					{
						"jwt": [
							"ADMIN"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "College registration details.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateCollegeDTO",
								"description": "College registration details."
							}
						}
					}
				}
			}
		},
...
		"/api/v1/colleges/{id}": {
			"get": {
				"operationId": "GetById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Get college details by ID.\n\n### Intent & Business Purpose\nFetches specific college profile details, contact information, location coordinates, and active status.\n\n### Path Parameters\n- `id` *(required string)*: College UUID.",
				"tags": [
					"Colleges"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "College UUID.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Update college profile details.\n\n### Intent & Business Purpose\nModifies college tenant information, address, contact details, or active status.\n\n### Target Audience & Roles\n- **Allowed Roles:** System Administrator (`ADMIN`).\n\n### Path Parameters\n- `id` *(required string)*: College UUID.",
				"tags": [
					"Colleges"
				],
				"security": [
					{
						"jwt": [
							"ADMIN"
						]
					}
				],
				"parameters": [
					{
						"description": "College UUID.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "Updates payload.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateCollegeDTO",
								"description": "Updates payload."
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Deactivate a college tenant.\n\n### Intent & Business Purpose\nSoft deletes / deactivates a college tenant from the system.\n\n### Target Audience & Roles\n- **Allowed Roles:** System Administrator (`ADMIN`).\n\n### Path Parameters\n- `id` *(required string)*: College UUID.",
				"tags": [
					"Colleges"
				],
				"security": [
					{
						"jwt": [
							"ADMIN"
						]
					}
				],
				"parameters": [
					{
						"description": "College UUID.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "http://localhost:5000",
			"description": "Development server"
		}
	]
}